+++ packages/create-plugin/CHANGELOG.md
@@ -1,3 +1,15 @@
+# v5.26.7 (Wed Sep 24 2025)
+
+#### 🐛 Bug Fix
+
+- Chore: shares the externals config between bundlers [#2141](https://github.com/grafana/plugin-tools/pull/2141) ([@hugohaggmark](https://github.com/hugohaggmark))
+
+#### Authors: 1
+
+- Hugo Häggmark ([@hugohaggmark](https://github.com/hugohaggmark))
+
+---
+
 # v5.26.6 (Tue Sep 23 2025)
 
 #### 🐛 Bug Fix